+++ lib/constants.dart
@@ -3,6 +3,7 @@ abstract class Constants {
   static const String generate = 'Generate';
   static const String project = 'Project';
   static const String build = 'Build';
+  static const String shorebird = 'Shorebird';
 
   static const String dev = 'dev';
   static const String dartDefine = '--dart-define';