+++ lib/project/get/get_command.dart
@@ -27,10 +27,6 @@ class GetCommand extends Command {
 
     final yaml = YamlHelper.loadFileYaml(argMorphemeYaml);
 
-    // if (yaml.containsKey('repository')) {
-    //   await 'morpheme repository --morpheme-yaml "$argMorphemeYaml"'.run;
-    // }
-
     if (argGenerateL10n) {
       await 'morpheme l10n --morpheme-yaml "$argMorphemeYaml"'.run;
     }